requirements.txt
yfinance
minio
confluent-kafka
pandas==1.3.5
pyspark
apache-airflow[spark]
apache-airflow-providers-apache-spark
news-please
lxml_html_clean
llama-index
llama-index-embeddings-openai
llama-index-vector-stores-chroma
# aiohttp 
# aiokafka
# llama-index-vector-stores-postgres
# elasticsearch
# llama-index-vector-stores-elasticsearch
# InstructorEmbedding
# transformers==4.20.0
# datasets>=2.2.0
# pyarrow==8.0.0
# jsonlines
# numpy
# requests>=2.26.0
# scikit_learn>=1.0.2
# scipy
# sentence_transformers>=2.2.0
# torch
# tqdm
# rich
# tensorboard
# huggingface-hub>=0.19.0
